// P1Frontend/src/interfaces/ReimbursementInterface.ts
import { UserInterface } from "./UserInterface";

export interface ReimbursementInterface {
    reimbursementId?: number,
    description: string,
    amount: number,
    status?: string,
    user?: UserInterface
}

//every car needs a registered make, model, and isFourWheelDrive
//carId and userId are optional - we don't supply these values if we're inserting a car
    //carId is autogenerated. userId will be supplied by the saved logged in user data